+++ BitFaster.Caching/Lru/AfterAccessPolicy.cs
@@ -3,19 +3,27 @@ using System.Runtime.CompilerServices;
 
 namespace BitFaster.Caching.Lru
 {
-    internal readonly struct DiscretePolicy<K, V> : IDiscreteItemPolicy<K, V>
+    /// <summary>
+    /// Implement an expire after access policy.
+    /// </summary>
+    internal readonly struct AfterAccessPolicy<K, V> : IItemPolicy<K, V, LongTickCountLruItem<K, V>>
     {
-        private readonly IExpiryCalculator<K, V> expiry;
+        private readonly Duration timeToLive;
         private readonly Time time;
 
-        public TimeSpan TimeToLive => TimeSpan.Zero;
+        ///<inheritdoc/>
+        public TimeSpan TimeToLive => timeToLive.ToTimeSpan();
 
-        public DiscretePolicy(IExpiryCalculator<K, V> expiry)
+        /// <summary>
+        /// Initializes a new instance of the AfterReadTickCount64Policy class with the specified time to live.
+        /// </summary>
+        /// <param name="timeToLive">The time to live.</param>
+        public AfterAccessPolicy(TimeSpan timeToLive)
         {
-            if (expiry == null)
-                Throw.ArgNull(ExceptionArgument.expiry);
+            if (timeToLive <= TimeSpan.Zero || timeToLive > Time.MaxRepresentable)
+                Throw.ArgOutOfRange(nameof(timeToLive), $"Value must greater than zero and less than {Time.MaxRepresentable}");
 
-            this.expiry = expiry;
+            this.timeToLive = Duration.FromTimeSpan(timeToLive);
             this.time = new Time();
         }
 
@@ -23,17 +31,14 @@ namespace BitFaster.Caching.Lru
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public LongTickCountLruItem<K, V> CreateItem(K key, V value)
         {
-            var expiry = this.expiry.GetExpireAfterCreate(key, value);
-            return new LongTickCountLruItem<K, V>(key, value, (expiry + Duration.SinceEpoch()).raw);
+            return new LongTickCountLruItem<K, V>(key, value, Duration.SinceEpoch().raw);
         }
 
         ///<inheritdoc/>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void Touch(LongTickCountLruItem<K, V> item)
         {
-            var currentExpiry = new Duration(item.TickCount - this.time.Last);
-            var newExpiry = expiry.GetExpireAfterRead(item.Key, item.Value, currentExpiry);
-            item.TickCount = this.time.Last + newExpiry.raw;
+            item.TickCount = this.time.Last;
             item.WasAccessed = true;
         }
 
@@ -41,10 +46,7 @@ namespace BitFaster.Caching.Lru
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void Update(LongTickCountLruItem<K, V> item)
         {
-            var time = Duration.SinceEpoch();
-            var currentExpiry = new Duration(item.TickCount) - time;
-            var newExpiry = expiry.GetExpireAfterUpdate(item.Key, item.Value, currentExpiry);
-            item.TickCount = (time + newExpiry).raw;
+            item.TickCount = Duration.SinceEpoch().raw;
         }
 
         ///<inheritdoc/>
@@ -52,7 +54,7 @@ namespace BitFaster.Caching.Lru
         public bool ShouldDiscard(LongTickCountLruItem<K, V> item)
         {
             this.time.Last = Duration.SinceEpoch().raw;
-            if (this.time.Last > item.TickCount)
+            if (this.time.Last - item.TickCount > this.timeToLive.raw)
             {
                 return true;
             }
